#306 ReactJs: Split main test suite into test file per react Component
